# services/business-agency-saas-api/public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 'Laravel API Documentation'
  description: ''
  version: 1.0.0
servers:
  -
    url: 'http://localhost'
...
components:
  securitySchemes:
    default:
      type: http
      scheme: bearer
      description: 'You can retrieve your token by visiting your dashboard and clicking <b>Generate API token</b>.'
security:
  -
    default: []
paths:
  /api/login:
    post:
      summary: ''
      operationId: postApiLogin
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  description: 'Must be a valid email address.'
                  example: gbailey@example.net
                password:
                  type: string
                  description: ''
                  example: '|]|{+-'
              required:
                - email
                - password
  /api/register:
    post:
      summary: ''
      operationId: postApiRegister
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                tenant_name:
                  type: string
                  description: 'Must not be greater than 255 characters.'
                  example: b
                tenant_domain:
                  type: string
                  description: 'Must not be greater than 255 characters.'
                  example: 'n'
                  nullable: true
                name:
                  type: string
                  description: 'Must not be greater than 255 characters.'
                  example: g
                email:
                  type: string
                  description: 'Must be a valid email address. Must not be greater than 255 characters.'
                  example: rowan.gulgowski@example.com
                password:
                  type: string
                  description: 'Must be at least 8 characters.'
                  example: 'BNvYgxwmi/#iw/kX'
              required:
                - tenant_name
                - name
                - email
                - password
  /api/ai-settings:
    post:
      summary: ''
      operationId: postApiAiSettings
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                ai_provider_default:
                  type: string
                  description: ''
                  example: gemini
                  enum:
                    - openai
                    - gemini
              required:
                - ai_provider_default
  /api/integrations:
    post:
      summary: "POST /api/integrations\nSecurely store the keys."
      operationId: pOSTapiintegrationsSecurelyStoreTheKeys
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                service:
                  type: string
                  description: ''
                  example: architecto
                value:
                  type: object
                  description: ''
                  example: []
                  properties: {  }
                key:
                  type: string
                  description: "Ensure 'value' is an array from frontend."
                  example: architecto
                  nullable: true
                is_active:
                  type: boolean
                  description: ''
                  example: false
                is_brain:
                  type: boolean
                  description: ''
                  example: false
              required:
                - service
                - value
                - is_active
                - is_brain
  '/api/integrations/{id}':
    put:
      summary: ''
      operationId: putApiIntegrationsId
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                service:
                  type: string
                  description: ''
                  example: architecto
                value:
                  type: object
                  description: ''
                  example: []
                  properties: {  }
                key:
                  type: string
                  description: "Ensure 'value' is an array from frontend."
                  example: architecto
                  nullable: true
                is_active:
                  type: boolean
                  description: ''
                  example: false
                is_brain:
                  type: boolean
                  description: ''
                  example: true
              required:
                - service
                - value
                - is_active
                - is_brain
    delete:
      summary: 'DELETE /api/integrations/{id}'
      operationId: dELETEapiintegrationsid
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
    parameters:
      -
        in: path
        name: id
        description: 'The ID of the integration.'
        example: architecto
        required: true
        schema:
          type: string
  /api/ai-agents:
    post:
      summary: 'Store a new agent.'
      operationId: storeANewAgent
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: 'Must not be greater than 255 characters.'
                  example: b
                slug:
                  type: string
                  description: 'Ensure slug is unique per tenant. Must contain only letters, numbers, dashes and underscores. Must not be greater than 255 characters.'
                  example: 'n'
                brain:
                  type: string
                  description: ''
                  example: architecto
                model:
                  type: string
                  description: "The service key (e.g. 'openai')."
                  example: architecto
                system_prompt:
                  type: string
                  description: "e.g. 'gpt-4'."
                  example: architecto
                  nullable: true
                user_prompt:
                  type: string
                  description: ''
                  example: architecto
                  nullable: true
                tools:
                  type: object
                  description: ''
                  example: null
                  properties: {  }
                  nullable: true
                triggers:
                  type: object
                  description: ''
                  example: null
                  properties: {  }
                  nullable: true
                is_active:
                  type: boolean
                  description: 'Logic to save triggers handled below or via relationship.'
                  example: true
              required:
                - name
                - slug
                - brain
                - model
...
  /api/ai-agent/settings:
    post:
      summary: ''
      operationId: postApiAiAgentSettings
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                ai_provider_default:
                  type: string
                  description: ''
                  example: gemini
                  enum:
                    - openai
                    - gemini
                    - claude
              required:
                - ai_provider_default
  /api/tenants:
    post:
      summary: ''
      operationId: postApiTenants
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: 'Must not be greater than 255 characters.'
                  example: b
                domain:
                  type: string
                  description: 'Must not be greater than 255 characters.'
                  example: 'n'
                  nullable: true
                status:
                  type: string
                  description: ''
                  example: suspended
                  enum:
                    - active
                    - suspended
                  nullable: true
                plan_id:
                  type: string
                  description: 'The <code>id</code> of an existing record in the plans table.'
                  example: architecto
                expires_at:
                  type: string
                  description: 'Must be a valid date.'
                  example: '2026-01-25T17:18:15'
                  nullable: true
                crm_config:
                  type: object
                  description: ''
                  example: null
                  properties:
                    entity_name_singular:
                      type: string
                      description: 'This field is required when <code>crm_config</code> is present.'
                      example: architecto
                    entity_name_plural:
                      type: string
                      description: 'This field is required when <code>crm_config</code> is present.'
                      example: architecto
                    statuses:
                      type: array
                      description: 'This field is required when <code>crm_config</code> is present. Must have at least 1 items.'
                      example: null
                      items:
                        type: object
                        properties:
                          slug:
                            type: string
                            description: 'This field is required when <code>crm_config</code> is present.'
                            example: architecto
                          label:
                            type: string
                            description: 'This field is required when <code>crm_config</code> is present.'
                            example: architecto
                          color:
                            type: string
                            description: 'This field is required when <code>crm_config</code> is present.'
                            example: architecto
                  nullable: true
              required:
                - name
                - plan_id
  '/api/tenants/{tenant_id}':
    patch:
      summary: ''
      operationId: patchApiTenantsTenant_id
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: 'Must not be greater than 255 characters.'
                  example: b
                domain:
                  type: string
                  description: 'Must not be greater than 255 characters.'
                  example: 'n'
                  nullable: true
                status:
                  type: string
                  description: ''
                  example: active
                  enum:
                    - active
                    - suspended
                plan_id:
                  type: string
                  description: 'The <code>id</code> of an existing record in the plans table.'
                  example: null
                expires_at:
                  type: string
                  description: 'Must be a valid date.'
                  example: '2026-01-25T17:18:15'
                  nullable: true
                crm_config:
                  type: object
                  description: ''
                  example: null
                  properties:
                    entity_name_singular:
                      type: string
                      description: 'This field is required when <code>crm_config</code> is present.'
                      example: architecto
                    entity_name_plural:
                      type: string
                      description: 'This field is required when <code>crm_config</code> is present.'
                      example: architecto
                    statuses:
                      type: array
                      description: 'This field is required when <code>crm_config</code> is present. Must have at least 1 items.'
                      example: null
                      items:
                        type: object
                        properties:
                          slug:
                            type: string
                            description: 'This field is required when <code>crm_config</code> is present.'
                            example: architecto
                          label:
                            type: string
                            description: 'This field is required when <code>crm_config</code> is present.'
                            example: architecto
                          color:
                            type: string
                            description: 'This field is required when <code>crm_config</code> is present.'
                            example: architecto
                  nullable: true
    delete:
      summary: ''
      operationId: deleteApiTenantsTenant_id
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
    parameters:
      -
        in: path
        name: tenant_id
        description: 'The ID of the tenant.'
        example: 16
        required: true
        schema:
          type: integer
  /api/tenants/crm-config:
    post:
      summary: ''
      operationId: postApiTenantsCrmConfig
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                entity_name_singular:
                  type: string
                  description: 'Must not be greater than 50 characters.'
                  example: b
                entity_name_plural:
                  type: string
                  description: 'Must not be greater than 50 characters.'
                  example: 'n'
                statuses:
                  type: array
                  description: 'Must have at least 1 items.'
                  example:
                    - []
                  items:
                    type: object
                    properties:
                      slug:
                        type: string
                        description: ''
                        example: architecto
                      label:
                        type: string
                        description: ''
                        example: architecto
                      color:
                        type: string
                        description: ''
                        example: architecto
                    required:
                      - slug
                      - label
                      - color
              required:
                - entity_name_singular
                - entity_name_plural
                - statuses
...
  '/api/webhooks/{id}':
    delete:
      summary: ''
      operationId: deleteApiWebhooksId
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
    parameters:
      -
        in: path
        name: id
        description: 'The ID of the webhook.'
        example: architecto
        required: true
        schema:
          type: string
...